Move SDK version 16 requirements into a class that's not instantiated unless the requirement is met
