+++ hyperion-geiger-counter/src/main/java/com/willowtreeapps/hyperion/geigercounter/GeigerCounterPlugin.java
@@ -7,6 +7,8 @@ import com.willowtreeapps.hyperion.plugin.v1.PluginModule;
 @AutoService(Plugin.class)
 public class GeigerCounterPlugin extends Plugin {
 
+    public static final String LOG_TAG = "Hyperion Geiger Counter";
+
     @Override
     public PluginModule createPluginModule() {
         return new GeigerCounterModule();
